requirements.txt
python-dotenv
requests
beautifulsoup4
selenium
pandas
pytz
PyMySQL
undetected-chromedriver
pandas
scikit-learn
sqlalchemy
xgboost
lightgbm
google_auth_oauthlib
google-api-python-client
google
jinja2
tenacity